Remove individual services from client (#33)
Method names are now generated "as is" from operation IDs. This means that it made the individual service types redundant.

The service types themselves only referred to the service which was the Client type itself. They were a bit redundant and not strictly necessary. Also manual step was needed for updating the Client which was easy to forget to add.