Beginning of path gen refactor (#55)
This commit modifies the paths.go generation functions to only write to file from a single place. Eventually templates will be used instead of forming the structure with several fmt.Printf(), a small testing function for this was added.

Additionally, the writeMethod() function has been separated into two to begin the work of having different functions for methods if necessary.