Remove outdated checks in Makefile and bump tools
We have a number of checks to remove `vendor` from the list of go
modules to operate on, but these are no longer present in the
repository. Likewise, we don't format `pb.go` and `.twirp.go` files,
which are no longer present.

Remove the redundant checks and bump our tools to their latest versions.