Make Profile and Token mutually exclusive
Error out if callers attempt to use both the `Token/Host` and
`Profile/UseDefaultProfile` options.

Also update tests to only set `$HOME` to our temp dir when requested.
This allows us to validate that custom config dirs are actually being
used and we're not just falling back on the default.