Add option to use credentials from CLI (#244)
The Oxide CLI stores authentication information in a well-known format.
For convenience, we should should allow SDK users to use these, rather
then requiring that it be manually set in the environment or directly
set.

Add functionality to `Config` to use Oxide config files, either with a
default profile, or with an explicitly provided profile name. We use the
files found in `$HOME/.config/oxide` by default. This can be can be
overridden with the `ConfigDir` field.

As with `Host`/`Token`, profiles will override the
`OXIDE_HOST`/`OXIDE_TOKEN` environment variables. `Profile` and
`UseDefaultProfile` are mutually exclusive with each other and the
`Host`/`Token` options to prevent confusion on the order or
precedence.