generate: handle map[string]any types correctly.
When an openapi field includes an `additionalProperties` key, it represents a
map to the type specified in the schema reference. We already correctly handle
the special case of mapping a string to a slice of the referenced type as a
`map[string][]type`, but not the case where the nested type isn't a slice. This
patch adds logic for the missing `map[string[type` case.