Handle required nullable types. (#329)
The omicron api includes fields that are nullable (i.e. may be
explicitly set to null) but required (i.e. they can't be omitted). To
handle these types in the sdk, we treat them as pointers, and don't set
`omitempty` in the struct tags.

This patch also updates omicron to latest to pull in newly added or
updated fields.