Add public MakeRequest method to client.
Most of the time, sdk users will interact with the api using generated client
methods. However, in some cases it may be useful to interact with the api more
directly, analogous to `oxide api` in the cli. For example, users may want to
invoke an api method that isn't yet supported in the sdk. This patch adds a new
public method called MakeRequest that allows users to send custom requests to
the api.