feat: allow custom json encoder/decoder.
The encoding/json package in the standard library isn't particularly
performant, and there have been many first-party and third-party efforts to
improve it. This patch allows users of the sdk to configure a custom json codec
to marshall and unmarshall json with their tools of choice. Note: we can update
to encoding/json/v2 as the default once it's marked stable in a future golang
release.