experiment: typed enums with interfaces.
Implement typed enums in go by changing the type from `any` to an interface.
The interface is defined by a single private method to ensure that exactly one
variaent of the enum is populated. We also implement a custom UnmarshalJSON to
choose the correct implementation to unmarshal into based on the discriminator
field.