Handle nullable array items. (#341)
Openapi represents an array of nullable items using an allOf with
exactly one reference entry (items.allOf[0].$ref), rather than the usual
items.$ref. We currently treat these types as any, which means users
have to cast or marshal and unmarshal data to the correct type. This
patch adds a check for this edge case and replaces the any with the
appropriate concrete type.

Part of #340.