Represent allOf of length 1 with concrete types.
We currently represent allOf types as `any` in go, since we haven't yet
implemented code generation logic to render types for unions of multiple
schemas. However, all allOf types in nexus have length of exactly 1; they're
used to work around limitations in openapi, and not to represent schema unions.
This patch adds an exception in createAllOf, so that we replace allOf types
with length 1 with their referenced schema. If we encounter an allOf with
multiple items, we'll continue to represent it as `any`, but this never happens
in nexus as of this writing.