Refactor and test enum type creation (#347)
In thinking through #342, I noticed that the enum type creation logic in
createOneOf is a bit hard to follow, and not tested thoroughly. In
advance of changing that behavior, I wanted to clean up the existing
logic and add some tests. This should make the code more readable and
easier to change in the future.

Relates to #350. 
Relates to SSE-123.